+++ bin/virtual_device.dart
@@ -8,10 +8,22 @@ import 'package:virtual_device/src/executable/stop_command.dart';
 import 'package:virtual_device/src/executable/wipe_command.dart';
 
 void main(List<String> args) async {
-  final runner = CommandRunner(
+  var runner = CommandRunner(
     'virtual_device',
     'Manage iOS simulators and Android emulators from the command line',
-  )
+  );
+  runner.argParser.addOption(
+    'platform',
+    allowed: ['ios', 'android'],
+    abbr: 'p',
+    defaultsTo: 'ios',
+    allowedHelp: {
+      'ios': 'Start an iOS device',
+      'android': 'Start an Android device'
+    },
+  );
+
+  runner = runner
     ..addCommand(CreateCommand())
     ..addCommand(DeleteCommand())
     ..addCommand(ListCommand())